python-ai/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Image Processing
albumentations>=1.3.0
scikit-image>=0.21.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
